+++ src/main/kotlin/kr/getx/fitnessteachers/entity/User.kt
@@ -9,7 +9,7 @@ data class User(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "UserID")
-    val userId: Long = 0L,
+    val userId: Int = 0,
     @Column(name = "User_Social_Media_ID")
     val userSocialMediaId: String? = null,
     @Column(name = "Username")
